Cambio de responsabilidades en contextos, notificaciones en CRUDS empleados y usuarios
