+++ src/components/DeleteModal.jsx
@@ -20,10 +20,10 @@ const DeleteModal = ({ onCancel,onConfirm, elements, message  }) => {
           </div>
         </div>
         <div className='h-20 overflow-y-scroll bg-slate-100 mt-2 p-2' >
-          {elements?.map((empl,indx) => {
-            return empl.isSelected ?
+          {elements?.map((elmt,indx) => {
+            return elmt.isSelected ?
               <p key={indx}>
-                {empl.nombre + " " + empl.apellidos}
+                {elmt.nombre + " " + elmt.apellidos}
               </p>
               : null
           })}
@@ -35,7 +35,7 @@ const DeleteModal = ({ onCancel,onConfirm, elements, message  }) => {
             Cancelar
           </button>
           <button
-            onClick={onConfirm}
+            onClick={()=>onConfirm(elements)}
             className='bg-rose-500 w-full py-1 ml-2 rose-opacity'>
             Continuar
           </button>
