+++ src/context/AppContext.jsx
@@ -2,6 +2,7 @@ import React, { useState } from "react";
 import { useEffect } from "react";
 import { useContext } from "react";
 import { toast, ToastContainer } from "react-toastify";
+import { useAuth } from "./AuthContext";
 
 const AppContext = React.createContext()
 
@@ -9,27 +10,264 @@ export function useApp() {
   return useContext(AppContext)
 }
 
+const apiEmpleadosUrl = 'http://127.0.0.1:8000/api/empleados/'
+const apiMaquinasUrl = 'http://127.0.0.1:8000/api/maquinas/'
+const apiEmpleadoMaquinaUrl = 'http://127.0.0.1:8000/api/empleados_maquina/'
+const apiEmpleadoMaquinasUrl = 'http://127.0.0.1:8000/api/empleado_maquinas/'
+const imageEndPoint = 'http://127.0.0.1:8000'
+
+const empleadosColumns = [
+  { name: 'Nombre', attribute: 'nombre' },
+  { name: 'Apellidos', attribute: 'apellidos' },
+  { name: 'Dirección', attribute: 'direccion' },
+  { name: 'Seguro Social', attribute: 'ns' },
+  { name: 'Teléfono', attribute: 'telefono' },
+  { name: 'Departamento', attribute: 'departamento' },
+]
+
+const maquinasColumns = [
+  { name: 'Número', attribute: 'numero' },
+  { name: 'Línea', attribute: 'linea' },
+  { name: 'Marca', attribute: 'marca' },
+  { name: 'Modelo', attribute: 'modelo' },
+  { name: 'Número de Serie', attribute: 'ns' },
+  { name: 'Fecha de Adquisición', attribute: 'fechaAdquisicion' },
+  { name: 'Otros', attribute: 'otros' },
+  { name: 'Departamento', attribute: 'departamento' },
+]
+
 export function AppProvider({ children }) {
 
-  const notify = (message, error = false ) => {
-    let options = {
-        position: "bottom-right",
-        autoClose: 5000,
-        hideProgressBar: false,
-        closeOnClick: true,
-        pauseOnHover: true,
-        draggable: true,
-        progress: undefined,
-        theme: "light",
+  const { session, notify } = useAuth()
+
+  const [fetchingEmpleados, setFetchingEmpleados] = useState(false)
+  const [allEmpleados, setAllEmpleados] = useState([])
+
+  const [fetchingMaquinas, setFetchingMaquinas] = useState(false)
+  const [allMaquinas, setAllMaquinas] = useState([])
+
+
+
+  
+
+  const getEmpleados = async () => {
+    setFetchingEmpleados(true)
+    let response = await fetch(apiEmpleadosUrl, {
+      method: 'GET',
+      headers: {
+        "Content-Type": "application/json",
+        'Authorization': 'Bearer ' + session.access
+      }
+    })
+    if (response.status == 200) {
+      let data = await response.json()
+      let formatData = data.map((empl) => ({
+        ...empl,
+        isSelected: false,
+        fotografia: empl.fotografia ? imageEndPoint + empl.fotografia : ''
+      }))
+      setAllEmpleados(formatData)
+      return formatData
     }
-    error ? toast.error(message,options) : toast.success(message,options)
-}
+    setFetchingEmpleados(false)
+  }
+
+  const saveEmpleado = async (values, objEmpleado, maquinas, isEdit) => {
+    
+    let formData = new FormData()
+    formData.append('nombre', values.nombre)
+    formData.append('apellidos', values.apellidos)
+    formData.append('direccion', values.direccion)
+    formData.append('telefono', values.telefono)
+    formData.append('correo', values.correo)
+    formData.append('ns', values.ns)
+    formData.append('usuario', values.usuario)
+    formData.append('contrasena', values.contrasena)
+    if ((objEmpleado.fotografia) instanceof File)
+      formData.append('fotografia', objEmpleado.fotografia)
+    formData.append('departamento', values.departamento)
+    formData.append('tipo', values.tipo)
+    
+    let maquinasIds = []
+    maquinas.forEach(m => maquinasIds.push({ id: m.idMaquina }))
+
+    if (!isEdit) {
+      //    Creacion de un Nuevo Empleado 
+      let response = await fetch(apiEmpleadosUrl, {
+        method: 'POST',
+        body: formData,
+        headers: { 'Authorization': 'Bearer ' + session.access }
+      })
+      //    Espero la respuesta para obtener el nuevo Id 
+      const { message, empleado } = await response.json()
+      notify(message)
+
+      if ( response.ok ) {
+        //    Asigno las Maquinas 
+        let response2 = await fetch(apiEmpleadoMaquinasUrl, {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+            'Authorization': 'Bearer ' + session.access
+          },
+          body: JSON.stringify({ idEmpleado: empleado.idEmpleado, maquinas: maquinasIds }),
+        })
+        let error = response2.ok ? false : true
+        let data = await response2.json()
+        notify(data.message, error)
+      }
+    }
+    else {
+      //    Actaulizo los datos del Empleado
+      console.log(values)
+      await fetch(apiEmpleadosUrl + values.idEmpleado, {
+        method: 'PUT',
+        body: formData,
+        headers: { 'Authorization': 'Bearer ' + session.access }
+      })
+        .then(response => response.json())
+        .then(data => notify(data.message))
+
+      if( maquinas.length === 0 ) return
+
+      //    Asigno Sus nuevas maquinas
+      let response = await fetch(apiEmpleadoMaquinasUrl, {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+          'Authorization': 'Bearer ' + session.access
+        },
+        body: JSON.stringify({ idEmpleado: values.idEmpleado, maquinas: maquinasIds }),
+      })
+      let data = await response.json()
+      notify(data.message)
+
+    }
+  }
+
+  const deleteEmpleados = async (listaEmpleados) => {
+    listaEmpleados.forEach(async (e) => {
+      if (e.isSelected) {
+        let response = await fetch(apiEmpleadosUrl + e.idEmpleado, {
+          method: 'DELETE',
+          headers: {
+            'Authorization': 'Bearer ' + session.access
+          }
+        })
+        console.log( response )
+        if( response.ok )
+        {
+          let data = await response.json()
+          console.log(data)
+          notify(data.message)
+        }else{
+          notify('Error al eliminar', true.message)
+        }
+      }
+    })
+  }
+
+  const getEmpleadoMaquinas = async (empId) => {
+    let response = await fetch(apiEmpleadoMaquinaUrl + empId, {
+      method: 'GET',
+      headers: {
+        "Content-Type": "application/json",
+        'Authorization': 'Bearer ' + session.access
+      }
+    })
+    if (response.status === 200) {
+      let assigned = await response.json()
+      return assigned
+    }
+    return []
+  }
+
+  const getMaquinas = async () => {
+    setFetchingMaquinas(true)
+    let response = await fetch(apiMaquinasUrl, {
+      method: 'GET',
+      headers: {
+        "Content-Type": "application/json",
+        'Authorization': 'Bearer ' + session.access
+      }
+    })
+    //console.log( 'getting maquinas ',response )
+    if (response.status === 200) {
+      let maquinas = await response.json()
+      setAllMaquinas(maquinas)
+      return maquinas
+    }
+    setFetchingMaquinas(false)
+  }
+
+  const saveMaquina = async (values, isEdit) => {
+
+    let formData = new FormData()
+    formData.append('numero', values.numero)
+    formData.append('linea', values.linea)
+    formData.append('marca', values.marca)
+    formData.append('modelo', values.modelo)
+    formData.append('ns', values.ns)
+    formData.append('fechaAdquisicion', values.fechaAdquisicion)
+    formData.append('otros', values.otros)
+    formData.append('departamento', values.departamento)
+
+    if (!isEdit) {
+      //Creacion de un nueva maquina 
+      await fetch(apiMaquinasUrl, {
+        method: 'POST',
+        body: formData,
+        headers: {
+          'Authorization': 'Bearer ' + session.access
+        }
+      })
+        .then(response => response.json())
+        .then(data => notify(data.message))
+
+    } else {
+      //Actualizando los datos de la maquina
+      await fetch(apiMaquinasUrl + values.idMaquina, {
+        method: 'PUT',
+        body: formData,
+        headers: {
+          'Authorization': 'Bearer ' + session.access
+        }
+      })
+        .then(response => response.json())
+        .then(data => alert(data.message))
+    }
+  }
+
+  const deleteMaquinas = async (listaMaquinas) => {
+    listaMaquinas.forEach(async (e) => {
+      if (e.isSelected) {
+        await fetch(apiMaquinasUrl + e.idMaquina, {
+          method: 'DELETE',
+          headers: {
+            'Authorization': 'Bearer ' + session.access
+          }
+        })
+          .then(response => response.json())
+          .then(data => alert('Maquinas Eliminadas:', data))
+      }
+    })
+  }
 
   return (
     <AppContext.Provider
       value={{
-       notify
+        fetchingEmpleados,
+        allEmpleados, getEmpleados, empleadosColumns,
+        saveEmpleado, deleteEmpleados,
+
+        fetchingMaquinas,
+        allMaquinas, getMaquinas, maquinasColumns,
+        saveMaquina, deleteMaquinas,
+
+        getEmpleadoMaquinas,
+        notify
       }}>
+
       {children}
       <ToastContainer position="bottom-right" />
     </AppContext.Provider>
